python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
netCDF4==1.6.5
h5py==3.10.0
geopandas==0.14.0
rasterio==1.3.9
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
folium==0.15.0
python-multipart==0.0.6
aiofiles==23.2.1
